Move testing classes to dinocpu.test.* heirarchy
- General testing classes (like 'CPUTesterDriver' and 'InstTests') were moved
  to dinocpu.test
- Component tests (like 'ALUControlUnitTest' and 'RegisterFileUnitTest') were
  moved to dinocpu.test.components
- Memory tests (like 'CombinMemoryUnitTest' and 'NonCombinMemoryUnitTest')
  were moved to dinocpu.test.memory
- Gradescope tests (the lab tests) were moved to dinocpu.test.grader
