+++ src/test/scala/components/ALUControlUnitTest.scala
@@ -1,12 +1,10 @@
 // Unit tests for the ALU control logic
 
-package dinocpu
+package dinocpu.test.components
 
 import chisel3._
-
-import chisel3.iotesters
 import chisel3.iotesters.{ChiselFlatSpec, Driver, PeekPokeTester}
-
+import dinocpu._
 
 class ALUControlUnitTester(c: ALUControl) extends PeekPokeTester(c) {
   private val ctl = c
