+++ src/test/scala/grading/Lab3Tests.scala
@@ -1,11 +1,15 @@
-package dinocpu
+// Lab 3 tester
 
-import chisel3.iotesters.Driver
+package dinocpu.test.grader
+
+import dinocpu._
+import dinocpu.test._
 
 import com.gradescope.jh61b.grader.{GradedTest,GradedTestRunnerJSON}
-import org.junit.Test;
+import org.junit.Test
 import org.scalatest.junit.JUnitSuite
 import org.junit.runner.RunWith
+
 
 @RunWith(classOf[GradedTestRunnerJSON])
 class Lab3Grader extends JUnitSuite {
