+++ src/test/scala/labs/Lab1Test.scala
@@ -5,8 +5,8 @@
 package dinocpu
 
 import chisel3._
-
 import chisel3.iotesters.{ChiselFlatSpec, Driver, PeekPokeTester}
+import dinocpu.test._
 
 
 class ALUControlUnitRTypeTester(c: ALUControl) extends PeekPokeTester(c) {
