+++ src/test/scala/labs/Lab2Test.scala
@@ -2,12 +2,13 @@
 // If you name your test class something that ends with "TesterLab2" it will
 // automatically be run when you use `Lab2 / test` at the sbt prompt.
 
-
 package dinocpu
 
-import chisel3._
+import chisel3.iotesters.{ChiselFlatSpec, Driver}
+import dinocpu.test._
+import dinocpu.test.components._
 
-import chisel3.iotesters.{ChiselFlatSpec, Driver, PeekPokeTester}
+
 /**
   * This is a trivial example of how to run this Specification
   * From within sbt use:
